refactor: move section flare to its own component, rename tags folder, clean up
