+++ packages/api-reference/src/components/SectionFlare/SectionFlare.vue
@@ -1,0 +1,21 @@
+<template>
+  <!-- For adding gradients + animations to introduction of documents that :before / :after won't work for -->
+  <div class="section-flare">
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+    <div class="section-flare-item"></div>
+  </div>
+</template>
+
+<style scoped>
+.section-flare {
+  top: 0;
+  right: 0;
+  pointer-events: none;
+}
+</style>
