+++ packages/api-reference/src/components/SectionFlare/index.ts
@@ -1,0 +1,1 @@
+export { default as SectionFlare } from './SectionFlare.vue'
